[FLINK-14441][Table-Common] Fix ValueLiteralExpression#getValueAs when valueClass is Period.class

+++ flink-table/flink-table-common/src/main/java/org/apache/flink/table/expressions/ValueLiteralExpression.java
@@ -104,7 +104,7 @@ public final class ValueLiteralExpression implements ResolvedExpression {
 			convertedValue = Duration.ofMillis(longVal);
 		}
 
-		else if (valueClass == Period.class && clazz == Integer.class) {
+		else if (valueClass == Period.class && clazz == Long.class) {
 			final Period period = (Period) value;
 			convertedValue = period.toTotalMonths();
 		}
